fix #327
