+++ raphtory-graphql/src/model/mod.rs
@@ -29,12 +29,14 @@ use std::{
     io::BufReader,
     ops::Deref,
 };
+use utils::path_prefix;
 use uuid::Uuid;
 
 pub(crate) mod algorithm;
 pub(crate) mod filters;
 pub(crate) mod graph;
 pub(crate) mod schema;
+pub(crate) mod utils;
 
 #[derive(Debug)]
 pub struct MissingGraph;
@@ -221,23 +223,6 @@ impl Mut {
             .to_string();
 
         if new_graph_name.ne(&graph_name) {
-            fn path_prefix(path: String) -> Result<String> {
-                let elements: Vec<&str> = path.split('/').collect();
-                let size = elements.len();
-                return if size > 2 {
-                    let delimiter = "/";
-                    let joined_string = elements
-                        .iter()
-                        .take(size - 1)
-                        .copied()
-                        .collect::<Vec<_>>()
-                        .join(delimiter);
-                    Ok(joined_string)
-                } else {
-                    Err("Invalid graph path".into())
-                };
-            }
-
             path = path_prefix(path)? + "/" + &Uuid::new_v4().hyphenated().to_string();
         }
 
@@ -277,6 +262,7 @@ impl Mut {
 
         new_subgraph.update_constant_properties([("lastUpdated", Prop::I64(timestamp * 1000))])?;
         new_subgraph.update_constant_properties([("uiProps", Prop::Str(props.into()))])?;
+        new_subgraph.update_constant_properties([("path", Prop::Str(path.clone().into()))])?;
 
         new_subgraph.save_to_file(path)?;
